requirements.txt
certifi==2018.8.24
chardet==3.0.4
clearbit==0.1.7
defusedxml==0.5.0
Django==1.11
django-allauth==0.37.1
django-rest-auth==0.9.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.7
oauthlib==2.1.0
Pillow==5.2.0
pyhunter==0.3
PyJWT==1.6.4
python-dotenv==0.9.1
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.23
